Change love-manager deployment configuration to split manager-producer into a list of managers.
This is inspired by how we did the deployment on the bare metal nodes.
Basically, the managers that receives data from the love-producers are split into a static number of managers.
Besides that, we deploy dedicated managers for the queue and m1m3.
The queue are critical systems to serve data to LOVE and therefore deserves a dedicated manager.
The M1M3 has a large throughtput and, as such, also deserves its own dedicated manager.